backend/requirements.txt
asgiref==3.11.1
Django==5.2.13
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
google-auth==2.49.2
PyJWT==2.12.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.1
